# api/service/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: carl.a.smigielski@saic.com
    name: Carl Smigielski
  description: An API for managing evaluations of Security Technical Implementation
    Guide (STIG) assessments.
  title: STIG Management API
  version: 1.0.0
servers:
- url: http://localhost:54000/stig-manager/api
paths:
...
  /users:
    get:
      operationId: getUsers
      parameters:
      - description: |
          Additional properties to include in the response.
        explode: true
        in: query
        name: projection
        required: false
        schema:
          items:
            enum:
            - stigAssets
            type: string
          type: array
        style: form
      - description: Elevate the user context for this request if user is permitted
          (canAdmin)
        explode: true
        in: query
        name: elevate
        required: false
        schema:
          type: boolean
        style: form
      - explode: true
        in: query
        name: role
        required: false
        schema:
          $ref: '#/components/schemas/UserRole'
        style: form
      - description: Selects Users mapped to a Package
        explode: true
        in: query
        name: packageId
        required: false
        schema:
          type: integer
        style: form
      - description: Selects Users mapped to a STIG
        explode: true
        in: query
        name: benchmarkId
        required: false
        schema:
          type: string
        style: form
      - description: Selects Users exactly matching a department string
        explode: true
        in: query
        name: dept
        required: false
        schema:
          type: string
        style: form
      - description: Selects Users matching the condition
        explode: true
        in: query
        name: canAdmin
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserProjected'
                type: array
          description: UserProjected response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:user:read
      summary: Return a list of Users accessible to the requester
      tags:
      - User
      x-swagger-router-controller: User
    post:
      operationId: createUser
      parameters:
      - description: |
          Additional properties to include in the response.
        explode: true
        in: query
        name: projection
        required: false
        schema:
          items:
            enum:
            - stigAssets
            type: string
          type: array
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAssign'
        required: true
        x-name: body
      responses:
        "201":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserProjected'
                type: array
          description: UserProjected response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:user
      summary: Create a User
      tags:
      - User
      x-codegen-request-body-name: body
      x-swagger-router-controller: User
...
components:
...
  schemas:
    Error:
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
      required:
      - code
      - message
      type: object
...
    AssetBasic:
      properties:
        assetId:
          type: integer
        name:
          type: string
        dept:
          type: string
      type: object
...
    UserRole:
      enum:
      - IA Staff
      - IA Officer
      - IA Workforce
      type: string
    User:
      properties:
        userId:
          readOnly: true
          type: integer
        username:
          type: string
        displayName:
          type: string
        dept:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        canAdmin:
          type: boolean
      type: object
    UserProjected:
      allOf:
      - $ref: '#/components/schemas/User'
      - $ref: '#/components/schemas/UserProjected_allOf'
...
    UserAssign:
      allOf:
      - $ref: '#/components/schemas/User'
      - $ref: '#/components/schemas/UserAssign_allOf'
...
    StigAssetAssigned:
      properties:
        saId:
          type: integer
        benchmarkId:
          type: string
        asset:
          $ref: '#/components/schemas/AssetBasic'
      type: object
...
    UserProjected_allOf:
      properties:
        stigAssets:
          items:
            $ref: '#/components/schemas/StigAssetAssigned'
          type: array
    UserAssign_allOf:
      properties:
        saIds:
          description: The STIG/Asset pairs (saIds) mapped to this Asset
          items:
            type: integer
          type: array
...
  securitySchemes:
    oauth:
      flows:
        implicit:
          authorizationUrl: http://localhost:8080/auth/realms/cmsat/protocol/openid-connect/auth
          scopes:
            stig-manager: Top-level scope
            stig-manager:stig: Full access to the STIG collection
            stig-manager:stig:read: Read access to the STIG collection
            stig-manager:package: Full access to the Package collection
            stig-manager:package:read: Read-only access to the Package collection
            stig-manager:user: Full access to the Users collection
            stig-manager:user:read: Read-only access to the Users collection
      type: oauth2